+++ src/main/java/emi/mtg/deckbuilder/view/MainWindow.java
@@ -333,7 +333,18 @@ public class MainWindow extends Stage {
 			fillMenu.getItems().add(fillZoneMenuItem);
 		}
 
-		menu.getItems().addAll(changePrintingMenuItem, tagsMenu, fillMenu);
+		MenuItem deleteImages = new MenuItem("Delete Saved Images");
+		deleteImages.setOnAction(ae -> {
+			menu.cards.forEach(ci -> {
+				try {
+					Context.get().images.deleteSavedImages(ci.printing());
+				} catch (IOException e) {
+					e.printStackTrace(); // These aren't fatal as far as I'm concerned...
+				}
+			});
+		});
+
+		menu.getItems().addAll(changePrintingMenuItem, tagsMenu, fillMenu, deleteImages);
 
 		return menu;
 	}